Create an Access Database
This script will create an Access Database. It will then create a table within, populate that table, and read it back  (and display it). 
It is interesting to note that you do not have to have Microsoft Access installed on the system for this script to  be able to run! 
Features: 
∙ Create an Access database 
∙ Create an Access database table 
∙ Populate an Access database table 
∙ Process an Access database/table